pyproject.toml
[tool.poetry]
name = "todoist-tui"
version = "0.1.0"
description = ""
authors = ["Miles Budden <contact@milesbudden.com>"]

[tool.poetry.dependencies]
python = "^3.7"
picotui = "^1.0.1"
requests = "^2.24.0"
todoist-python = "^8.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
